Move `allocate()` js runtime function to runtime_legacy.js.
We no longer use this function internally.  Move it to a new
file called `runtime_legacy.js` which will be the home for such
function until we are sure there are no more external users.
